requirements.txt
streamlit
python-dotenv
google-generativeai
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
langchain-community